+++ CHANGELOG.md
@@ -65,11 +65,11 @@
 
 * wip... WHEN RELEASE UPDATE `pyproject.toml`, `pdm build_doc` & `pdm build_all`
 
-* tbd... invoice customer search for different keys, like: "reg com", "CUI", "bank / IBAN / cont", ... (area saved in `_area_to_search`)
+* tbd... invoice customer search for other keys: "reg com", "bank / IBAN / cont", ... (area saved in `_area_to_search`)
 
+* tbd... invoice customer search and persist for "RegistrationName"
 
-
-* WIP... 2401??piu_? invoice customer search for "CUI"  #NOTE: rdinv() lines 255 - 261
+* 240105piu_b invoice customer search and persist for "CUI"
 
 * 240105piu_a `rdinv.def get_excel_data_at_label(...)` changed strategy for DOWN search-method made it optional with default True (useful for Partners set-of KVs where is supposed to be or IN-LABEL or in RIGHT but NOT DOWN because there is a list of KVs not just one placed anywhere in Excel doc)  #TODO tgis is subject of doc update